edit(build): Added about page and updated image cloud
