+++ about/admin.py
@@ -1,0 +1,22 @@
+from django.contrib import admin
+from .models import About, CollaborateRequest
+from django_summernote.admin import SummernoteModelAdmin
+
+
+@admin.register(About)
+class AboutAdmin(SummernoteModelAdmin):
+    summernote_fields = ('content',)
+
+# Note: admin.ModelAdmin is the standard way of registering
+#       our model with the admin panel. We do it differently
+#       above because we are supplying Summernote fields.
+#       If you want to customise the admin panel view in your
+#       own projects, then inherit from admin.ModelAdmin like
+#       we do below.
+
+@admin.register(CollaborateRequest)
+class CollaborateRequestAdmin(admin.ModelAdmin):
+
+    list_display = ('message', 'read',)
+
+admin.site.register(Comment)
