+++ about/forms.py
@@ -1,0 +1,8 @@
+from .models import CollaborateRequest
+from django import forms
+
+
+class CollaborateForm(forms.ModelForm):
+    class Meta:
+        model = CollaborateRequest
+        fields = ('name', 'email', 'message')
