+++ about/templates/about/about.html
@@ -2,11 +2,18 @@
 {% load static %}
 {% load crispy_forms_tags %}
 
+
 {% block content %}
 <div class="container mt-5">
     <div class="row">
         <div class="col-12 col-md-4 text-center">
-            <img src="{% static 'images/nobody.jpg' %}" width="75%">
+            {% if "placeholder" in about.profile_image.url %}
+            <img src="{% static 'images/nobody.jpg' %}" width="75%"
+                alt="placeholder image" >
+            {% else %}
+            <img src="{{ about.profile_image.url }}" width="75%"
+                alt="{{ about.title }}">
+            {% endif %}
         </div>
         <div class="col-12 col-md-8">
             <h2>{{ about.title }}</h2>
@@ -30,13 +37,13 @@
                 and we can get the ball rolling!</p>
 
             <!-- add your form here. Your submit button should 
-            have the classes of btn, btn-secondary -->
+                                have the classes of btn, btn-secondary -->
             <form method="post">
                 {{ collaborate_form | crispy }}
                 {% csrf_token %}
                 <button class="btn btn-secondary"
                     type="submit">Submit</button>
-            </form>
+                </form>
         </div>
     </div>
 </div>
