+++ blog/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Post, Comment, About, CollaborateRequest
+from .models import Post, Comment
 from django_summernote.admin import SummernoteModelAdmin
 
 @admin.register(Post) #this is a decorator
@@ -13,20 +13,5 @@ class PostAdmin(SummernoteModelAdmin):
 
 # Register your models here.
 
-@admin.register(About)
-class AboutAdmin(SummernoteModelAdmin):
-    summernote_fields = ('content',)
 
-# Note: admin.ModelAdmin is the standard way of registering
-#       our model with the admin panel. We do it differently
-#       above because we are supplying Summernote fields.
-#       If you want to customise the admin panel view in your
-#       own projects, then inherit from admin.ModelAdmin like
-#       we do below.
 
-@admin.register(CollaborateRequest)
-class CollaborateRequestAdmin(admin.ModelAdmin):
-
-    list_display = ('message', 'read',)
-
-admin.site.register(Comment)
