+++ blog/forms.py
@@ -1,4 +1,4 @@
-from .models import Comment, CollaborateRequest
+from .models import Comment
 from django import forms
 
 
@@ -6,8 +6,3 @@ class CommentForm(forms.ModelForm):
     class Meta:
         model = Comment
         fields = ('body',)
-
-class CollaborateForm(forms.ModelForm):
-    class Meta:
-        model = CollaborateRequest
-        fields = ('name', 'email', 'message')
