+++ blog/models.py
@@ -37,22 +37,3 @@ class Comment(models.Model):
         ordering = ["-created_on"]
     def __str__(self):
         return f"Comment by {self.author} on{self.post}"
-
-
-class About(models.Model):
-    title = models.CharField(max_length=200)
-    updated_on = models.DateTimeField(auto_now=True)
-    content = models.TextField()
-
-    def __str__(self):
-        return self.title
-
-
-class CollaborateRequest(models.Model):
-    name = models.CharField(max_length=200)
-    email = models.EmailField()
-    message = models.TextField()
-    read = models.BooleanField(default=False)
-
-    def __str__(self):
-        return f"Collaboration request from {self.name}"
