+++ codestar/settings.py
@@ -26,6 +26,11 @@ TEMPLATES_DIR = os.path.join(BASE_DIR, 'templates')
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = 'django-insecure-n-y2z*xk+mc$^uv73y!z+6l8wv%2_*i=s_13c^7a3a&1b!vm!_'
 
+# if DEBUG:
+#     SECRET_KEY = 'mysecretkey'
+# else:
+#     SECRET_KEY = os.environ['SECRET_KEY']
+
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
